+++ tests/testthat/test-assertions.R
@@ -29,16 +29,10 @@ test_that("assert_has_variables Test 2: no error if a required variable exists",
 # assert_filter_cond ----
 ## Test 3: `assert_filter_cond` works as expected ----
 test_that("assert_filter_cond Test 3: `assert_filter_cond` works as expected", {
-  fc <- quo(AGE == 64)
+  fc <- expr(AGE == 64)
   expect_identical(
     assert_filter_cond(fc),
     fc
-  )
-
-  fc <- quo()
-  expect_error(
-    assert_filter_cond(arg = fc),
-    "Argument `fc` is missing, with no default"
   )
 
   expect_identical(
@@ -46,10 +40,17 @@ test_that("assert_filter_cond Test 3: `assert_filter_cond` works as expected", {
     fc
   )
 
-  fc <- quo("string")
+  fc <- expr("string")
   expect_error(
     assert_filter_cond(arg = fc),
     "`fc` must be a filter condition but is `\"string\"`"
+  )
+
+  vals <- c("A", "B")
+  fc <- expr(VAR %in% !!vals)
+  expect_identical(
+    assert_filter_cond(arg = fc),
+    fc
   )
 })
 
@@ -57,7 +58,7 @@ test_that("assert_filter_cond Test 3: `assert_filter_cond` works as expected", {
 ## Test 4: error if not a dataframe ----
 test_that("assert_data_frame Test 4: error if not a dataframe", {
   example_fun <- function(dataset) {
-    assert_data_frame(dataset, required_vars = vars(STUDYID, USUBJID))
+    assert_data_frame(dataset, required_vars = exprs(STUDYID, USUBJID))
   }
   expect_error(
     example_fun(c(1, 2, 3))
@@ -67,7 +68,7 @@ test_that("assert_data_frame Test 4: error if not a dataframe", {
 ## Test 5: error if dataframe is grouped ----
 test_that("assert_data_frame Test 5: error if dataframe is grouped", {
   example_fun <- function(dataset) {
-    assert_data_frame(dataset, required_vars = vars(STUDYID, USUBJID))
+    assert_data_frame(dataset, required_vars = exprs(STUDYID, USUBJID))
   }
 
   data <- tibble::tribble(
@@ -113,9 +114,9 @@ test_that("assert_character_scalar Test 7: error if input is a vector", {
 # assert_vars ----
 ## Test 8: no error if expected input ----
 test_that("assert_vars Test 8: no error if expected input", {
-  expect_invisible(assert_vars(vars(USUBJID, PARAMCD)))
+  expect_invisible(assert_vars(exprs(USUBJID, PARAMCD)))
   expect_invisible(assert_vars(
-    vars(APERSDT = APxxSDT, APEREDT = APxxEDT),
+    exprs(APERSDT = APxxSDT, APEREDT = APxxEDT),
     expect_names = TRUE
   ))
 })
@@ -123,17 +124,17 @@ test_that("assert_vars Test 8: no error if expected input", {
 ## Test 9: error if unexpected input ----
 test_that("assert_vars Test 9: error if unexpected input", {
   expect_error(assert_vars(AVAL + 1))
-  expect_error(assert_vars(rlang::exprs(USUBJID, PARAMCD)))
+  expect_error(assert_vars(rlang::quos(USUBJID, PARAMCD)))
   expect_error(assert_vars(c("USUBJID", "PARAMCD", "VISIT")))
-  expect_error(assert_vars(vars(USUBJID, AVAL + 2)))
-  expect_error(assert_vars(vars(APERSDT = APxxSDT, APxxEDT), expect_names = TRUE))
+  expect_error(assert_vars(exprs(USUBJID, AVAL + 2)))
+  expect_error(assert_vars(exprs(APERSDT = APxxSDT, APxxEDT), expect_names = TRUE))
 })
 
 # assert_data_frame ----
 ## Test 10: no error if optional is TRUE and `arg` is NULL ----
 test_that("assert_data_frame Test 10: no error if optional is TRUE and `arg` is NULL", {
   example_fun <- function(dataset) {
-    assert_data_frame(dataset, required_vars = vars(STUDYID, USUBJID), optional = TRUE)
+    assert_data_frame(dataset, required_vars = exprs(STUDYID, USUBJID), optional = TRUE)
   }
 
   expect_invisible(
@@ -144,7 +145,7 @@ test_that("assert_data_frame Test 10: no error if optional is TRUE and `arg` is 
 ## Test 11: error if required variables are missing ----
 test_that("assert_data_frame Test 11: error if required variables are missing", {
   example_fun <- function(dataset) {
-    assert_data_frame(dataset, required_vars = vars(STUDYID, USUBJID))
+    assert_data_frame(dataset, required_vars = exprs(STUDYID, USUBJID))
   }
 
   admiral_dm <- admiral.test::admiral_dm %>% select(-c(STUDYID, USUBJID))
@@ -157,7 +158,7 @@ test_that("assert_data_frame Test 11: error if required variables are missing", 
 ## Test 12: error if required variable is missing ----
 test_that("assert_data_frame Test 12: error if required variable is missing", {
   example_fun <- function(dataset) {
-    assert_data_frame(dataset, required_vars = vars(STUDYID, USUBJID))
+    assert_data_frame(dataset, required_vars = exprs(STUDYID, USUBJID))
   }
 
   admiral_dm <- admiral.test::admiral_dm %>% select(-c(USUBJID))
@@ -323,7 +324,7 @@ test_that("assert_logical_scalar Test 19: error if `arg` is not TRUE or FALSE", 
 ## Test 20: no error if optional = TRUE and `arg` = NULL ----
 test_that("assert_symbol Test 20: no error if optional = TRUE and `arg` = NULL", {
   f <- function(var) {
-    v <- enquo(var)
+    v <- enexpr(var)
   }
 
   example_fun <- function(arg) {
@@ -340,7 +341,7 @@ test_that("assert_symbol Test 20: no error if optional = TRUE and `arg` = NULL",
 ## Test 21: `assert_symbol` throws an error if `arg` is missing ----
 test_that("assert_symbol Test 21: `assert_symbol` throws an error if `arg` is missing", {
   f <- function(var) {
-    v <- enquo(var)
+    v <- enexpr(var)
   }
 
   example_fun <- function(arg) {
@@ -357,7 +358,7 @@ test_that("assert_symbol Test 21: `assert_symbol` throws an error if `arg` is mi
 ## Test 22: `assert_symbol` throws an error if `arg` is not a symbol ----
 test_that("assert_symbol Test 22: `assert_symbol` throws an error if `arg` is not a symbol", {
   f <- function(var) {
-    v <- enquo(var)
+    v <- enexpr(var)
   }
 
   example_fun <- function(arg) {
@@ -374,7 +375,7 @@ test_that("assert_symbol Test 22: `assert_symbol` throws an error if `arg` is no
 ## Test 23: `assert_symbol` does not throw an error if `arg` is a symbol ----
 test_that("assert_symbol Test 23: `assert_symbol` does not throw an error if `arg` is a symbol", {
   f <- function(var) {
-    v <- enquo(var)
+    v <- enexpr(var)
   }
 
   admiral_dm <- admiral.test::admiral_dm
@@ -394,7 +395,7 @@ test_that("assert_symbol Test 23: `assert_symbol` does not throw an error if `ar
 ## Test 24: `assert_expr` does not throw an error if `arg` is an expression ----
 test_that("assert_expr Test 24: `assert_expr` does not throw an error if `arg` is an expression", {
   f <- function(var) {
-    v <- enquo(var)
+    v <- enexpr(var)
   }
 
   example_fun <- function(arg) {
@@ -411,7 +412,7 @@ test_that("assert_expr Test 24: `assert_expr` does not throw an error if `arg` i
 ## Test 25: no error if optional is TRUE and `arg` is NULL ----
 test_that("assert_expr Test 25: no error if optional is TRUE and `arg` is NULL", {
   f <- function(var) {
-    v <- enquo(var)
+    v <- enexpr(var)
   }
 
   example_fun <- function(arg) {
@@ -428,7 +429,7 @@ test_that("assert_expr Test 25: no error if optional is TRUE and `arg` is NULL",
 ## Test 26: `assert_expr` throws an error if `arg` is missing ----
 test_that("assert_expr Test 26: `assert_expr` throws an error if `arg` is missing", {
   f <- function(var) {
-    v <- enquo(var)
+    v <- enexpr(var)
   }
 
   example_fun <- function(arg) {
@@ -443,7 +444,7 @@ test_that("assert_expr Test 26: `assert_expr` throws an error if `arg` is missin
 ## Test 27: `assert_expr` throws an error if `arg` is not an expression ----
 test_that("assert_expr Test 27: `assert_expr` throws an error if `arg` is not an expression", {
   f <- function(var) {
-    v <- enquo(var)
+    v <- enexpr(var)
   }
 
   example_fun <- function(arg) {
@@ -479,7 +480,7 @@ test_that("assert_vars Test 29: error if some elements of `arg` are not unquoted
   }
 
   expect_error(
-    example_fun(vars(USUBJID, PARAMCD, NULL))
+    example_fun(exprs(USUBJID, PARAMCD, NULL))
   )
 })
 
@@ -897,7 +898,7 @@ test_that("assert_varval_list Test 61: error if `arg` is not a list of var-value
   }
 
   expect_error(
-    example_fun(vars(USUBJID, PARAMCD, NULL))
+    example_fun(exprs(USUBJID, PARAMCD, NULL))
   )
 })
 
@@ -908,7 +909,7 @@ test_that("assert_varval_list Test 62: error if `required_elements` are missing 
   }
 
   expect_error(
-    example_fun(vars(DTHSEQ = AESEQ))
+    example_fun(exprs(DTHSEQ = AESEQ))
   )
 })
 
@@ -930,7 +931,7 @@ test_that("assert_varval_list Test 64: error if `accept_expr` is TRUE and value 
   }
 
   expect_error(
-    example_fun(vars(DTHSEQ = TRUE))
+    example_fun(exprs(DTHSEQ = TRUE))
   )
 })
 
@@ -941,7 +942,7 @@ test_that("assert_varval_list Test 65: error if `accept_expr` is FALSE and value
   }
 
   expect_error(
-    example_fun(vars(DTHSEQ = vars()))
+    example_fun(exprs(DTHSEQ = exprs()))
   )
 })
 
@@ -952,7 +953,7 @@ test_that("assert_varval_list Test 66: no error if an argument is a variable-val
   }
 
   expect_invisible(
-    example_fun(vars(DTHDOM = "AE", DTHSEQ = AESEQ))
+    example_fun(exprs(DTHDOM = "AE", DTHSEQ = AESEQ))
   )
 })
 
@@ -960,11 +961,13 @@ test_that("assert_varval_list Test 66: no error if an argument is a variable-val
 ## Test 67: no error if the elements fulfill a certain condition ----
 test_that("assert_list_element Test 67: no error if the elements fulfill a certain condition", {
   expect_invisible(
-    assert_list_element(vars(DTHDOM = "AE", DTHSEQ = AESEQ), "DTHSEQ", TRUE, message_text = "")
-  )
-  expect_invisible(
-    assert_list_element(vars(DTHDOM = "AE", DTHSEQ = admiral.test::admiral_dm), "DTHSEQ",
-      (admiral.test::admiral_dm)$DOMAIN == "DM",
+    assert_list_element(
+      list(
+        list(var = expr(DTHDT), val = 1),
+        list(var = expr(EOSDT), val = 0)
+      ),
+      element = "val",
+      condition = val >= 0,
       message_text = ""
     )
   )
@@ -973,10 +976,17 @@ test_that("assert_list_element Test 67: no error if the elements fulfill a certa
 ## Test 68: error if the elements do not fulfill the condition ----
 test_that("assert_list_element Test 68: error if the elements do not fulfill the condition", {
   expect_error(
-    assert_list_element(vars(DTHDOM = "AE", DTHSEQ = admiral.test::admiral_dm), "DTHSEQ",
-      (admiral.test::admiral_dm)$DOMAIN == "GM",
-      message_text = ""
-    )
+    assert_list_element(
+      input <- list(
+        list(var = expr(DTHDT), val = 1),
+        list(var = expr(EOSDT), val = -1)
+      ),
+      element = "val",
+      condition = val >= 0,
+      message_text = "Invalid value for `val`:"
+    ),
+    "Invalid value for `val`:\ninput[[2]]$val = -1",
+    fixed = TRUE
   )
 })
 
@@ -984,14 +994,14 @@ test_that("assert_list_element Test 68: error if the elements do not fulfill the
 ## Test 69: error if there is a one to many mapping ----
 test_that("assert_one_to_one Test 69: error if there is a one to many mapping", {
   expect_error(
-    assert_one_to_one(admiral.test::admiral_dm, vars(DOMAIN), vars(USUBJID))
+    assert_one_to_one(admiral.test::admiral_dm, exprs(DOMAIN), exprs(USUBJID))
   )
 })
 
 ## Test 70: error if there is a many to one mapping ----
 test_that("assert_one_to_one Test 70: error if there is a many to one mapping", {
   expect_error(
-    assert_one_to_one(admiral.test::admiral_dm, vars(USUBJID), vars(DOMAIN))
+    assert_one_to_one(admiral.test::admiral_dm, exprs(USUBJID), exprs(DOMAIN))
   )
 })
 
@@ -999,7 +1009,7 @@ test_that("assert_one_to_one Test 70: error if there is a many to one mapping", 
 ## Test 71: error if variable is not a date or datetime variable ----
 test_that("assert_date_var Test 71: error if variable is not a date or datetime variable", {
   example_fun <- function(dataset, var) {
-    var <- assert_symbol(enquo(var))
+    var <- assert_symbol(enexpr(var))
     assert_date_var(dataset = dataset, var = !!var)
   }